refactor: 컴포넌트 관련 데이터 폴더 위치 변경
