+++ src/components/filterBar/componentsData/SelectCategoryComponentData.ts
@@ -1,4 +1,4 @@
-import { MODAL_RESTAURANT_CATEGORY_UI_OPTIONS } from '../../selectOptions/restaurantCategoryUiOptions';
+import { MODAL_RESTAURANT_CATEGORY_UI_OPTIONS } from '../../../constants/selectOptions/restaurantCategoryUiOptions';
 
 const SELECT_CATEGORY_COMPONENT_DATA = Object.freeze({
   UI_OPTIONS: MODAL_RESTAURANT_CATEGORY_UI_OPTIONS,
