+++ src/components/filterBar/componentsData/SelectCategoryLabelComponentData.ts
