+++ src/components/filterBar/componentsData/SortByNameOrDistanceData.ts
