+++ src/components/filterBar/renderHandlers.ts
@@ -1,5 +1,5 @@
-import RESTAURANT_CATEGORY_DATA from '../../constants/componentOptions/selectCategory/RestaurantCategoryData';
-import SORT_BY_NAME_OR_CATEGORY_DATA from '../../constants/componentOptions/selectCategory/SortByNameOrDistanceData';
+import RESTAURANT_CATEGORY_DATA from './componentsData/RestaurantCategoryData';
+import SORT_BY_NAME_OR_CATEGORY_DATA from './componentsData/SortByNameOrDistanceData';
 import generateSelectElement from '../../uiUtils/generateSelectComponent';
 import convertHTMLStringToDOM from '../../utils/convertHTMLStringToDOM';
