+++ src/components/modal/inputDescription/renderHandlers.ts
@@ -1,6 +1,6 @@
-import DESCRIPTION_LABEL_COMPONENT_DATA from '../../../constants/componentOptions/description/DescriptionLabelComponentData';
-import DESCRIPTION_SPAN_COMPONENT_DATA from '../../../constants/componentOptions/description/DescriptionSpanComponentData';
-import DESCRIPTION_TEXT_AREA_COMPONENT_DATA from '../../../constants/componentOptions/description/DescriptionTextAreaComponentData';
+import DESCRIPTION_LABEL_COMPONENT_DATA from './componentsData/DescriptionLabelComponentData';
+import DESCRIPTION_SPAN_COMPONENT_DATA from './componentsData/DescriptionSpanComponentData';
+import DESCRIPTION_TEXT_AREA_COMPONENT_DATA from './componentsData/DescriptionTextAreaComponentData';
 import generateLabelComponent from '../../../uiUtils/generateLabelComponent';
 import generateSpanComponent from '../../../uiUtils/generateSpanComponent';
 import generateTextAreaComponent from '../../../uiUtils/generateTextAreaComponent';
