+++ src/components/modal/inputLink/componentsData/LinkLabelComponentData.ts
