+++ src/components/modal/inputLink/componentsData/LinkSpanComponentData.ts
