+++ src/components/modal/inputLink/renderHandlers.ts
@@ -1,6 +1,6 @@
-import LINK_INPUT_COMPONENT_DATA from '../../../constants/componentOptions/link/LinkInputComponentData';
-import LINK_LABEL_COMPONENT_DATA from '../../../constants/componentOptions/link/LinkLabelComponentData';
-import LINK_SPAN_COMPONENT_DATA from '../../../constants/componentOptions/link/LinkSpanComponentData';
+import LINK_INPUT_COMPONENT_DATA from './componentsData/LinkInputComponentData';
+import LINK_LABEL_COMPONENT_DATA from './componentsData/LinkLabelComponentData';
+import LINK_SPAN_COMPONENT_DATA from './componentsData/LinkSpanComponentData';
 import generateInputComponent from '../../../uiUtils/generateInputComponent';
 import generateLabelComponent from '../../../uiUtils/generateLabelComponent';
 import generateSpanComponent from '../../../uiUtils/generateSpanComponent';
