+++ src/components/modal/inputName/componentsData/NameInputComponentData.ts
