+++ src/components/modal/inputName/componentsData/NameLabelComponentData.ts
