+++ src/components/modal/inputName/renderHandlers.ts
@@ -1,5 +1,5 @@
-import NAME_INPUT_COMPONENT_DATA from '../../../constants/componentOptions/name/NameInputComponentData';
-import NAME_LABEL_COMPONENT_DATA from '../../../constants/componentOptions/name/NameLabelComponentData';
+import NAME_INPUT_COMPONENT_DATA from './componentsData/NameInputComponentData';
+import NAME_LABEL_COMPONENT_DATA from './componentsData/NameLabelComponentData';
 import generateInputComponent from '../../../uiUtils/generateInputComponent';
 import generateLabelComponent from '../../../uiUtils/generateLabelComponent';
 import convertHTMLStringToDOM from '../../../utils/convertHTMLStringToDOM';
