+++ src/components/modal/modalButton/componentsData/CancelButtonComponentData.ts
