+++ src/components/modal/selectDistance/componentsData/DistanceLabelComponentData.ts
